+++ backend/src/main.rs
@@ -9,9 +9,18 @@ pub mod log {
     tonic::include_proto!("log.v1");
 }
 
+pub mod dt {
+    tonic::include_proto!("dt.v1");
+}
+
 use log::{
     log_collector_service_server::{LogCollectorService, LogCollectorServiceServer},
     LogRequest, LogResponse,
+};
+
+use dt::{
+    dt_service_server::{DtService, DtServiceServer},
+    SyncRequest, SyncResponse,
 };
 
 #[derive(Debug, Default)]
@@ -28,21 +37,42 @@ impl LogCollectorService for LogCollector {
     }
 }
 
+#[derive(Debug, Default)]
+pub struct Dt {}
+
+#[tonic::async_trait]
+impl DtService for Dt {
+    async fn sync(&self, request: Request<SyncRequest>) -> Result<Response<SyncResponse>, Status> {
+        for event in request.get_ref().events.iter() {
+            println!("Event: {} {} {}", event.user_token, event.action, event.payload);
+        }
+
+        let reply = SyncResponse { success: true };
+
+        Ok(Response::new(reply))
+    }
+}
+
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
     let addr = "[::1]:80".parse()?;
     let log_service = LogCollector::default();
+    let dt_service = Dt::default();
 
     println!("WebService listening on {addr}");
 
-    let grpc_service = LogCollectorServiceServer::new(log_service);
-    let grpc_web_service = tonic_web::enable(grpc_service);
+    let log_grpc_service = LogCollectorServiceServer::new(log_service);
+    let log_grpc_web_service = tonic_web::enable(log_grpc_service);
+
+    let dt_grpc_service = DtServiceServer::new(dt_service);
+    let dt_grpc_web_service = tonic_web::enable(dt_grpc_service);
 
     let static_files_service = any_service(ServeDir::new("../frontend/dist"));
     let media_files_service = any_service(ServeDir::new("media"));
 
     let app = Router::new()
-        .nest_service("/grpc", grpc_web_service)
+        .nest_service("/grpc/log", log_grpc_web_service)
+        .nest_service("/grpc/dt", dt_grpc_web_service)
         .nest_service("/media", media_files_service)
         .fallback(static_files_service)
         .layer(